Combined the switch statements with the method
I noticed that the discount from the yearly quarters were not being added to the final cost, I found out it was because you couldn't return any values from the switch statements unless it was in it's own method. The program works fine now and the final discount is added to the final cost properly now.